// FrontEnd/src/app/core/models/store.model.ts
export interface Store {
  id: string;
  name: string;
  logoUrl?: string;
  phone?: string;
  email?: string;
  address?: string;
  primaryColor?: string;
  secondaryColor?: string;
  facebookUrl?: string;
  instagramUrl?: string;
  tiktokUrl?: string;
  whatsappUrl?: string;
  schedule?: string;
  createdAt?: string;
  updatedAt?: string;
}